Move ApiClientBase and ApiServiceErrors to core/api.
